git-svn-id: http://192.168.20.11/sequoiadb/misc/ci@6423 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ ant-sdbtask/src/com/sequoiadb/ant/sdbtask/SdbKillMasterNode.java
@@ -13,15 +13,21 @@ import com.sequoiadb.exception.BaseException;
 
 public class SdbKillMasterNode extends Task {
 	
-	private String uuid = null;
+	//private String uuid = null;
 	private String rgName = null;
+	private String localHostName = null ; 
 	private boolean failonerror = true;
 	private Process pro;
 
-	public void setSdbhandle(String value)
+	public void setLocalHostName ( String value )
 	{
-		uuid = value;
+		this.localHostName = value ; 
 	}
+	public String getLocalHostName ()
+	{
+		return this.localHostName ; 
+	}
+	
 	
 	public void setRgname(String value)
 	{
@@ -34,30 +40,35 @@ public class SdbKillMasterNode extends Task {
 	}
 	
 	public void execute() {
-		Object obj = this.getProject().getReference(uuid);
+		//Object obj = this.getProject().getReference(uuid);
+		/*
 		if (! (obj instanceof Sequoiadb))
 		{
 			throw new BuildException("The SdbUUID" + uuid + " cannot get Sequoiadb Object.");			
 		}
-		
+		*/
 		try
 		{
-			Sequoiadb sdb = (Sequoiadb) obj;
+			//Sequoiadb sdb = (Sequoiadb) obj;
+			Sequoiadb sdb = new Sequoiadb( this.getLocalHostName() ,50000 , "" ,"") ;
 			
-			int masterNodePort = sdb.getReplicaGroup(rgName).getMaster().getPort();
-			String shell = "ps -ef | grep sequoiadb\\(" + masterNodePort +" | awk '{print $2}'";
-			
-			try {
-				pro = Runtime.getRuntime().exec(new String[]{"sh" , "-c" , shell});
-			} catch (IOException e) {
-				e.printStackTrace();
-			}
-			BufferedReader br = new BufferedReader(new InputStreamReader( pro.getInputStream() ) ) ;
-			try {
-				String killpro = "kill -9 " + br.readLine();
-				Runtime.getRuntime().exec(new String[]{"sh","-c",killpro});
-			} catch (IOException e) {
-				e.printStackTrace();
+			if( this.getLocalHostName() == sdb.getReplicaGroup(this.rgName).getMaster().getNodeName().toString() )
+			{
+				int masterNodePort = sdb.getReplicaGroup(rgName).getMaster().getPort();
+				String shell = "ps -ef | grep sequoiadb\\(" + masterNodePort +" | awk '{print $2}'";
+				
+				try {
+					pro = Runtime.getRuntime().exec(new String[]{"sh" , "-c" , shell});
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+				BufferedReader br = new BufferedReader(new InputStreamReader( pro.getInputStream() ) ) ;
+				try {
+					String killpro = "kill -9 " + br.readLine();
+					Runtime.getRuntime().exec(new String[]{"sh","-c",killpro});
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
 			}
 		}
 		catch(BaseException e)
